line-bot/requirements.txt
line-bot-sdk
flask
paho-mqtt

#pip install -r requirements.txt
